SearchIndexArrayList
There is an array having unique elements is given as input. Find the index of an element k if present in the array, else -1.
Input
There is an integer n is given(size of array nums).
In second line, n space seperated integers are given.
in third line, an integer k is given.
1 <= n <= 105
0 <= nums[i] <= 106
0 <= i <= n-1
Output
Find the index of an element k if present in the array, else -1.
Example
Sample Input:
5
1 4 3 2 5
3

Sample Output:
2

Explanation: Index of element 3 is 2. So the answer will be 2.